Update index.php
Altered some variable names, added switch block, and made converted output display above the form on the same page, so the user can continue to enter temperatures to convert without refreshing the page.  Also added function to ensure the user enters or selects all of the required fields.  Failing to do so will prompt the user to try again.